// bank-account/src/types/user.ts
export interface UserParams {
  readonly fullName: string;
};

export interface UserBankAccount {
  readonly id: string;
  readonly userId: string;
  readonly bankCode: string;
  readonly accountBank: string;
  readonly accountNumber: string;
};

export default interface User {
  readonly id: string;
  readonly fullName: string;
};